[scroll-animations] Build fixes for including Styleable on ScrollTimeline https://bugs.webkit.org/show_bug.cgi?id=286637 rdar://143775363
Reviewed by Tim Nguyen.

Including Styleable on ScrollTimeline led to a header cycle, so remove the offending
headers on Styleable and fix the resulting unified build issues.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/animation/AnimationEffect.cpp:
* Source/WebCore/animation/AnimationTimelinesController.cpp:
(WebCore::AnimationTimelinesController::registerNamedScrollTimeline):
* Source/WebCore/animation/AnimationTimelinesController.h:
* Source/WebCore/animation/CSSAnimation.cpp:
(WebCore::CSSAnimation::syncStyleOriginatedTimeline):
* Source/WebCore/animation/KeyframeEffect.cpp:
* Source/WebCore/animation/KeyframeEffect.h:
* Source/WebCore/animation/ScrollTimeline.h:
* Source/WebCore/animation/ViewTimeline.h:
* Source/WebCore/animation/WebAnimation.h:
* Source/WebCore/animation/WebAnimationTypes.h:
* Source/WebCore/html/canvas/OffscreenCanvasRenderingContext2D.cpp:
* Source/WebCore/inspector/agents/InspectorAnimationAgent.cpp:
* Source/WebCore/platform/animation/AcceleratedEffect.cpp:
* Source/WebCore/platform/animation/Animation.h:
* Source/WebCore/style/Styleable.cpp:
(WebCore::Styleable::applyKeyframeEffects const):
* Source/WebCore/style/Styleable.h:
(WebCore::WeakStyleable::WeakStyleable):
(WebCore::WeakStyleable::element const):
(WebCore::WeakStyleable::pseudoElementIdentifier const):
(WebCore::Styleable::applyKeyframeEffects const): Deleted.

Canonical link: https://commits.webkit.org/289478@main
